cleanup: remove useSelectAll
