+++ prism-management-console-web/src/components/groupEditing/AddContactsModal/AddContactsModal.jsx
@@ -8,10 +8,14 @@ import ConnectionsFilter from '../../connections/Molecules/filter/ConnectionsFil
 import CustomButton from '../../common/Atoms/CustomButton/CustomButton';
 import { useCurrentGroupStore } from '../../../hooks/useGroupStore';
 import SelectAllButton from '../../newCredential/Molecules/RecipientsTable/SelectAllButton';
-import { useSelectAll } from '../../../hooks/useSelectAll';
 import { CONTACT_ID_KEY } from '../../../helpers/constants';
 import { filterByMultipleKeys } from '../../../helpers/filterHelpers';
 import './_style.scss';
+
+const getCheckedAndIndeterminateProps = (rows, selectedRows) => ({
+  checked: rows.length && selectedRows.length === rows.length,
+  indeterminate: Boolean(selectedRows.length && selectedRows.length !== rows.length)
+});
 
 const AddContactsModal = observer(({ visible, onCancel, onConfirm }) => {
   const { t } = useTranslation();
@@ -40,16 +44,26 @@ const AddContactsModal = observer(({ visible, onCancel, onConfirm }) => {
     setFilteredContacts(textFilter ? applyFilters(contactsNotInGroup) : contactsNotInGroup);
   }, [contactsNotInGroup, textFilter]);
 
-  const { loadingSelection, checkboxProps } = useSelectAll({
-    displayedEntities: filteredContacts,
-    entitiesFetcher: () => filteredContacts,
-    entityKey: CONTACT_ID_KEY,
-    selectedEntities: selectedContactIds,
-    setSelectedEntities: setSelectedContactIds,
-    isFetching: isLoadingContactsNotInGroup
-  });
+  const handleSelectAll = ev => {
+    const { checked } = ev.target;
+    setSelectedContactIds(checked ? filteredContacts.map(c => c[CONTACT_ID_KEY]) : []);
+  };
+
+  const handleContactSelect = (contactRecord, selected) => {
+    setSelectedContactIds(prevSelectedContactIds =>
+      selected
+        ? [...prevSelectedContactIds, contactRecord[CONTACT_ID_KEY]]
+        : prevSelectedContactIds.filter(cId => cId !== contactRecord[CONTACT_ID_KEY])
+    );
+  };
 
   const handleConfirm = () => onConfirm(selectedContactIds);
+
+  const checkboxProps = {
+    ...getCheckedAndIndeterminateProps(filteredContacts, selectedContactIds),
+    disabled: isLoadingContactsNotInGroup || !filteredContacts.length,
+    onChange: handleSelectAll
+  };
 
   const confirmButton = (
     <Row>
@@ -78,7 +92,7 @@ const AddContactsModal = observer(({ visible, onCancel, onConfirm }) => {
       <Row type="flex" align="middle" className="mb-3">
         <Col span={5}>
           <SelectAllButton
-            isLoadingSelection={loadingSelection}
+            isLoadingSelection={false}
             selectedEntities={selectedContactIds}
             checkboxProps={checkboxProps}
           />
@@ -101,7 +115,7 @@ const AddContactsModal = observer(({ visible, onCancel, onConfirm }) => {
             hasFiltersApplied
             isLoading={isLoadingContactsNotInGroup}
             selectedContactIds={selectedContactIds}
-            setSelectedContactIds={setSelectedContactIds}
+            onSelect={handleContactSelect}
             size="md"
           />
         </Col>
